requirements.txt
tensorflow==2.15.0
flask
numpy
matplotlib
scikit-learn
pillow